+++ packages/components/pie-toggle-switch/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @justeattakeaway/pie-toggle-switch
 
+## 0.14.0
+
+### Minor Changes
+
+- [Added] - @tagname jsdoc comment to top of component class and use new defineCustomElement function to define the components ([#905](https://github.com/justeattakeaway/pie/pull/905)) by [@jamieomaguire](https://github.com/jamieomaguire)
+
 ## 0.13.0
 
 ### Minor Changes